chore: Refactor directory structure pt 2 - fix filepaths
Fix filepath errors caused by refactoring code into new directories
