+++ app/components/SideBar/index.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import styles from './styles.css';
 // import { Button } from 'react-bootstrap';
-import AddContactButton from '../../containers/AddContactButton';
+import AddContactButton from '../../containers/Contacts/AddContactButton';
 class SideBar extends React.Component { // eslint-disable-line react/prefer-stateless-function
   render() {
     return (
