+++ app/containers/Contacts/ContactCard/styles.css
@@ -143,10 +143,10 @@ textarea{
   -webkit-transition: box-shadow 0.3s, border 0.3s;
 }
 .formField {
-  border: solid 1px #707070;
-  box-shadow: 0 0 5px 1px #969696;
-  -moz-box-shadow: 0 0 5px 1px #969696;
-  -webkit-box-shadow: 0 0 5px 1px #969696;
+  /*border: solid 1px #707070;*/
+  /*box-shadow: 0 0 5px 1px #969696;*/
+  /*-moz-box-shadow: 0 0 5px 1px #969696;*/
+  /*-webkit-box-shadow: 0 0 5px 1px #969696;*/
 }
 .formFieldError,
 .formFieldError:focus {
